Split all widgets into a structure under uix module following the Kivy pattern.
Made default colors for some widgets gray instead of blue.
