+++ flat_kivy/data/ui_elements.kv
@@ -77,7 +77,7 @@
         font_size: root.font_size
 
 
-<RaisedWidget>:
+<RaisedStyle>:
     canvas.before:
         BorderImage:
             border: self.border
@@ -331,7 +331,7 @@
     font_size: (self.size[0] + self.size[1]/2.)*self.scale
 
 
-<CheckBoxListItem>:
+<FlatCheckBoxListItem>:
     orientation: 'horizontal'
     active: checkbox.active
     ripple_color: app.get_color(self.ripple_color_tuple)
