+++ flat_kivy/uix/flatoptionbutton.py
@@ -1,0 +1,8 @@
+
+from kivy.properties import StringProperty
+
+from flat_kivy.uix.flattogglebutton import FlatToggleButton
+
+
+class FlatOptionButton(FlatToggleButton):
+    key = StringProperty(None)
